backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
dnspython==2.7.0
fastapi==0.109.2
h11==0.16.0
idna==3.10
pydantic==2.6.1
pydantic_core==2.33.2
pymongo==4.12.1
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.27.1
motor==3.3.2
pydantic-settings==2.1.0
pydantic[email]==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
